rename generic opintopolku-spring-security package. package name in rest test.
