+++ opintopolku-spring-security/src/main/java/fi/vm/sade/fi/vm/sade/javautils/opintopolku_spring_security/Authorizer.java
@@ -1,4 +1,4 @@
-package fi.vm.sade.authentication.business.service;
+package fi.vm.sade.fi.vm.sade.javautils.opintopolku_spring_security;
 
 import fi.vm.sade.authorization.NotAuthorizedException;
