+++ opintopolku-spring-security/src/main/java/fi/vm/sade/fi/vm/sade/javautils/opintopolku_spring_security/OrganisationHierarchyAuthorizer.java
@@ -1,6 +1,7 @@
-package fi.vm.sade.security;
+package fi.vm.sade.fi.vm.sade.javautils.opintopolku_spring_security;
 
 import fi.vm.sade.authorization.NotAuthorizedException;
+import fi.vm.sade.security.SimpleCache;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
